Unifying multi and single thread game versions. Seems ok except for with the loading. Need to find where multiple copies are interferring.
