--- src/test/java/game/GameSingleThreadTest.java
@@ -1,136 +1,0 @@
-package game;
-
-import actions.Action;
-import actions.ActionQueue;
-import org.junit.Assert;
-import org.junit.Test;
-
-/**
- * Make sure that the single-threaded and multi-threadable versions of the game match perfectly.
- */
-public class GameSingleThreadTest{
-
-    @Test
-    public void stepGame() {
-        GameThreadSafe gameMulti = new GameThreadSafe();
-        GameSingleThread gameSingle = GameSingleThread.getInstance();
-        gameSingle.makeNewWorld();
-
-        for (int i = 0; i < 4; i++) {
-            ActionQueue actionQueue = getSampleActions();
-            while (!actionQueue.isEmpty()) {
-                boolean[] command = actionQueue.pollCommand();
-                gameMulti.step(command);
-                gameSingle.step(command[0], command[1], command[2], command[3]);
-
-                float[] initStateMulti = gameMulti.getCurrentState().flattenState();
-                float[] initStateSingle = gameSingle.getCurrentState().flattenState();
-                Assert.assertArrayEquals(initStateMulti, initStateSingle, 1e-12f);
-            }
-
-            // Make sure that creating a new world does not affect the results.
-            gameMulti.makeNewWorld();
-            gameSingle.makeNewWorld();
-        }
-        gameSingle.releaseGame();
-    }
-
-    @Test
-    public void getInitialState() {
-        float[] initStateMulti = GameThreadSafe.getInitialState().flattenState();
-        float[] initStateSingle = GameSingleThread.getInitialState().flattenState();
-
-        Assert.assertArrayEquals(initStateMulti, initStateSingle, 1e-12f);
-    }
-
-    @Test
-    public void restoreFullState() {
-        GameSingleThread gameSingle = GameSingleThread.getInstance();
-        gameSingle.makeNewWorld();
-
-        // Run through the full queue with no saving/loading
-        ActionQueue actions = getSampleActions();
-        while (!actions.isEmpty()) {
-            gameSingle.step(actions.pollCommand());
-        }
-        State stateEndNoLoad = gameSingle.getCurrentState();
-
-        // Redo with save/load in the middle.
-        gameSingle.makeNewWorld();
-        actions = getSampleActions();
-        for (int i = 0; i < 30; i++) {
-            gameSingle.step(actions.pollCommand());
-        }
-
-        State stateBeforeLoad = gameSingle.getCurrentState();
-
-        // Save
-        byte[] fullState = gameSingle.getFullState();
-
-        // Step forward arbitrarily.
-        for (int i = 0; i < 10; i++) {
-            gameSingle.step(true, false, false, true);
-        }
-
-        // Load
-        gameSingle = gameSingle.restoreFullState(fullState);
-        State stateAfterLoad = gameSingle.getCurrentState();
-
-        // Make sure states at save and after load are equal.
-        Assert.assertArrayEquals(stateBeforeLoad.flattenState(), stateAfterLoad.flattenState(), 1e-15f);
-
-        // Finish the queue on the loaded game.
-        while (!actions.isEmpty()) {
-            gameSingle.step(actions.pollCommand());
-        }
-
-        State stateEndAfterLoad = gameSingle.getCurrentState();
-
-        Assert.assertArrayEquals(stateEndNoLoad.flattenState(), stateEndAfterLoad.flattenState(), 1e-15f);
-
-        gameSingle.releaseGame();
-    }
-
-    private static ActionQueue getSampleActions() {
-        // Ran MAIN_Search_LongRun to get these. 19 steps.
-        ActionQueue actionQueue = new ActionQueue();
-        actionQueue.addAction(new Action(1, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(34, new boolean[]{false, true, true, false}));
-        actionQueue.addAction(new Action(19, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(45, new boolean[]{true, false, false, true}));
-
-        actionQueue.addAction(new Action(10, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(38, new boolean[]{false, true, true, false}));
-        actionQueue.addAction(new Action(5, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(31, new boolean[]{true, false, false, true}));
-
-        actionQueue.addAction(new Action(21, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(21, new boolean[]{false, true, true, false}));
-        actionQueue.addAction(new Action(14, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(35, new boolean[]{true, false, false, true}));
-
-        actionQueue.addAction(new Action(10, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(23, new boolean[]{false, true, true, false}));
-        actionQueue.addAction(new Action(20, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(23, new boolean[]{true, false, false, true}));
-
-        actionQueue.addAction(new Action(13, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(20, new boolean[]{false, true, true, false}));
-        actionQueue.addAction(new Action(24, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(22, new boolean[]{true, false, false, true}));
-
-        actionQueue.addAction(new Action(18, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(23, new boolean[]{false, true, true, false}));
-        actionQueue.addAction(new Action(20, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(24, new boolean[]{true, false, false, true}));
-
-        actionQueue.addAction(new Action(21, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(20, new boolean[]{false, true, true, false}));
-        actionQueue.addAction(new Action(19, new boolean[]{false, false, false, false}));
-        actionQueue.addAction(new Action(21, new boolean[]{true, false, false, true}));
-
-        actionQueue.addAction(new Action(16, new boolean[]{false, false, false, false}));
-
-        return actionQueue;
-    }
-}
